🚀 update: 보호소 회원가입 step 에서 새로고침 시 /register/shelter로 이동하도록 로직 수정
