+++ src/app/register/shelter/[...slug]/Description.tsx
@@ -5,8 +5,8 @@ import EmphasizedTitle, {
 import TextArea from '@/components/common/TextField/TextArea';
 import useHeader from '@/hooks/useHeader';
 import { FieldValues, SubmitHandler, useFormContext } from 'react-hook-form';
-import { OnNextProps } from '../page';
-import * as styles from './../styles.css';
+import { OnNextProps } from './page';
+import * as styles from '../styles.css';
 import { useEffect } from 'react';
 
 export default function Description({ onSubmit }: OnNextProps) {
